Did the initial implementation of the results page
Before doing anything I found there was some unnecessary padding in for the footer, so I removed that. I also made sure that when search was clicked it would take you to the search results page. The main part of this is that I implemented the results/search result page and made it responsive.